+++ Interop/Runtime/src/native/kotlin/kotlinx/cinterop/NativeUtils.kt
@@ -20,8 +20,10 @@ import kotlin.native.internal.Intrinsic
 import kotlin.native.internal.TypedIntrinsic
 import kotlin.native.internal.IntrinsicType
 
-internal fun decodeFromUtf8(bytes: ByteArray): String = bytes.decodeToString()
 internal fun encodeToUtf8(str: String): ByteArray = str.encodeToByteArray()
+
+@SymbolName("Kotlin_CString_toKStringFromUtf8Impl")
+internal external fun CPointer<ByteVar>.toKStringFromUtf8Impl(): String
 
 @TypedIntrinsic(IntrinsicType.INTEROP_BITS_TO_FLOAT)
 external fun bitsToFloat(bits: Int): Float
