+++ runtime/src/main/cpp/Interop.cpp
@@ -17,8 +17,10 @@
 #include <assert.h>
 #include <stdio.h>
 #include <stdint.h>
+#include <string.h>
 
 #include "Alloc.h"
+#include "KString.h"
 #include "Memory.h"
 #include "MemorySharedRefs.hpp"
 #include "Types.h"
@@ -40,6 +42,10 @@ void Kotlin_Interop_disposeStablePointer(KNativePtr pointer) {
 OBJ_GETTER(Kotlin_Interop_derefStablePointer, KNativePtr pointer) {
   KRefSharedHolder* holder = reinterpret_cast<KRefSharedHolder*>(pointer);
   RETURN_OBJ(holder->ref<ErrorPolicy::kThrow>());
+}
+
+OBJ_GETTER(Kotlin_CString_toKStringFromUtf8Impl, const char* cstring) {
+  RETURN_RESULT_OF(StringFromUtf8Buffer, cstring, strlen(cstring));
 }
 
 }
