+++ runtime/src/main/cpp/KString.h
@@ -31,6 +31,8 @@ OBJ_GETTER(CreateStringFromUtf8, const char* utf8, uint32_t lengthBytes);
 char* CreateCStringFromString(KConstRef kstring);
 void DisposeCString(char* cstring);
 
+OBJ_GETTER(StringFromUtf8Buffer, const char* start, size_t size);
+
 #ifdef __cplusplus
 }
 #endif
